refactor: update API endpoint, enforce integer ratings, disable half-star input, and restructure offline sync payload

+++ lib/features/my_reports/providers/my_reports_provider.dart
@@ -59,7 +59,7 @@ class MyReportsNotifier extends AutoDisposeAsyncNotifier<List<Issue>> {
     await ApiClient.instance.dio.post(
       '/api/issues/$issueId/feedback',
       data: {
-        'rating': rating,
+        'rating': rating.toInt(),
         'comment': text,
       },
     );
